Extract shared semver functions
