+++ packages/cli-utils/src/commands/doctor/runner.ts
@@ -1,11 +1,11 @@
 import fs from 'node:fs/promises';
 import path from 'node:path';
-import semiver from 'semiver';
 
 import type { GraphQLSPConfig, LoadConfigResult } from '@gql.tada/internal';
 import { loadRef, loadConfig, parseConfig } from '@gql.tada/internal';
 
 import type { ComposeInput } from '../../term';
+import { MINIMUM_VERSIONS, semverComply } from '../../utils/semver';
 import { findGraphQLConfig } from './helpers/graphqlConfig';
 import * as vscode from './helpers/vscode';
 import * as logger from './logger';
@@ -22,11 +22,6 @@ const delay = (ms = 700) => {
   }
 };
 
-export const semiverComply = (version: string, compare: string) => {
-  const match = version.match(/\d+\.\d+\.\d+/);
-  return match ? semiver(match[0], compare) >= 0 : false;
-};
-
 const enum Messages {
   TITLE = 'Doctor',
   DESCRIPTION = 'Detects problems with your setup',
@@ -36,13 +31,6 @@ const enum Messages {
   CHECK_VSCODE = 'Checking VSCode setup',
   CHECK_SCHEMA = 'Checking schema',
 }
-
-export const MINIMUM_VERSIONS = {
-  typescript_embed_lsp: '5.5.0',
-  typescript: '4.1.0',
-  tada: '1.0.0',
-  lsp: '1.0.0',
-};
 
 export async function* run(): AsyncIterable<ComposeInput> {
   yield logger.title(Messages.TITLE, Messages.DESCRIPTION);
@@ -80,7 +68,7 @@ export async function* run(): AsyncIterable<ComposeInput> {
       `A version of ${logger.code('typescript')} was not found in your dependencies.\n` +
         logger.hint(`Is ${logger.code('typescript')} installed in this package?`)
     );
-  } else if (!semiverComply(typeScriptVersion[1], MINIMUM_VERSIONS.typescript)) {
+  } else if (!semverComply(typeScriptVersion[1], MINIMUM_VERSIONS.typescript)) {
     // TypeScript version lower than v4.1 which is when they introduced template lits
     yield logger.failedTask(Messages.CHECK_TS_VERSION);
     throw logger.errorMessage(
@@ -95,7 +83,7 @@ export async function* run(): AsyncIterable<ComposeInput> {
   yield logger.runningTask(Messages.CHECK_DEPENDENCIES);
   await delay();
 
-  const supportsEmbeddedLsp = semiverComply(
+  const supportsEmbeddedLsp = semverComply(
     typeScriptVersion[1],
     MINIMUM_VERSIONS.typescript_embed_lsp
   );
@@ -107,7 +95,7 @@ export async function* run(): AsyncIterable<ComposeInput> {
         `A version of ${logger.code('@0no-co/graphqlsp')} was not found in your dependencies.\n` +
           logger.hint(`Is ${logger.code('@0no-co/graphqlsp')} installed?`)
       );
-    } else if (!semiverComply(gqlspVersion[1], MINIMUM_VERSIONS.lsp)) {
+    } else if (!semverComply(gqlspVersion[1], MINIMUM_VERSIONS.lsp)) {
       yield logger.failedTask(Messages.CHECK_DEPENDENCIES);
       throw logger.errorMessage(
         `The version of ${logger.code(
@@ -127,7 +115,7 @@ export async function* run(): AsyncIterable<ComposeInput> {
       `A version of ${logger.code('gql.tada')} was not found in your dependencies.\n` +
         logger.hint(`Is ${logger.code('gql.tada')} installed?`)
     );
-  } else if (!semiverComply(gqlTadaVersion[1], '1.0.0')) {
+  } else if (!semverComply(gqlTadaVersion[1], '1.0.0')) {
     yield logger.failedTask(Messages.CHECK_DEPENDENCIES);
     throw logger.errorMessage(
       `The version of ${logger.code('gql.tada')} in your dependencies is out of date.\n` +
